fix(backend): corrige hallazgos reales de Sonar Quality Gate (ticket 006)
new_violations=5, verificado vía sonarqube-db:
- java:S2184 (x2): suma int+int pasada a un double (Vec4) sin cast
  explícito en AlphaAutoPackStrategy.faceAt -- cast a double antes de sumar.
- java:S5838 (x3): assertThat(...).isEqualTo(0) en vez de isZero() en los
  tests nuevos de AutoUv.

+++ backend/src/main/java/com/galgothstudio/backend/domain/uv/AlphaAutoPackStrategy.java
@@ -137,7 +137,7 @@ public final class AlphaAutoPackStrategy implements UvLayoutStrategy {
 	}
 
 	private static Face faceAt(int u0, int v0, int width, int height) {
-		return new Face(new Vec4(u0, v0, u0 + width, v0 + height), SINGLE_TEXTURE_INDEX);
+		return new Face(new Vec4(u0, v0, (double) u0 + width, (double) v0 + height), SINGLE_TEXTURE_INDEX);
 	}
 
 	private static Face faceOf(CuboidFaces faces, FaceName name) {
